Implement Alarm monitor
Add definition of alarm monitor in VNFD TOSCA template definition,
support it in VNFM plugin. It can integrate with scaling feature.

An WSGI filter is used to receive ceilometer action URL call and
translate the call into inner action, such as scalling.

The configuration group alarm_auth in tacker.conf is used as the
credentials for ceilometer action URL call. they must have 'admin'
role to query all related stuff for the action.

Implements blueprint: #alarm-based-monitoring-driver

+++ tacker/vnfm/monitor.py
@@ -44,7 +44,9 @@ CONF.register_opts(OPTS, group='monitor')
 
 
 def config_opts():
-    return [('monitor', OPTS), ('tacker', VNFMonitor.OPTS)]
+    return [('monitor', OPTS),
+            ('tacker', VNFMonitor.OPTS),
+            ('tacker', VNFAlarmMonitor.OPTS), ]
 
 
 def _log_monitor_events(context, vnf_dict, evt_details):
@@ -196,6 +198,72 @@ class VNFMonitor(object):
                             vnf=vnf_dict, kwargs=kwargs)
 
 
+class VNFAlarmMonitor(object):
+    """VNF Alarm monitor"""
+    OPTS = [
+        cfg.ListOpt(
+            'alarm_monitor_driver', default=['ceilometer'],
+            help=_('Alarm monitoring driver to communicate with '
+                   'Hosting VNF/logical service '
+                   'instance tacker plugin will use')),
+    ]
+    cfg.CONF.register_opts(OPTS, 'tacker')
+
+    # get alarm here
+    def __init__(self):
+        self._alarm_monitor_manager = driver_manager.DriverManager(
+            'tacker.tacker.alarm_monitor.drivers',
+            cfg.CONF.tacker.alarm_monitor_driver)
+
+    def update_vnf_with_alarm(self, vnf, policy_name, policy_dict):
+        params = dict()
+        params['vnf_id'] = vnf['id']
+        params['mon_policy_name'] = policy_name
+        _log_monitor_events(t_context.get_admin_context(),
+                            vnf,
+                            "update vnf with alarm")
+        driver = policy_dict['triggers']['resize_compute'][
+            'event_type']['implementation']
+        policy_action = policy_dict['triggers']['resize_compute'].get('action')
+        if not policy_action:
+            return
+        alarm_action_name = policy_action['resize_compute'].get('action_name')
+        if not alarm_action_name:
+            return
+        params['mon_policy_action'] = alarm_action_name
+        alarm_url = self.call_alarm_url(driver, vnf, params)
+        _log_monitor_events(t_context.get_admin_context(),
+                            vnf,
+                            "Alarm url invoked")
+        return alarm_url
+        # vnf['attribute']['alarm_url'] = alarm_url ---> create
+        # by plugin or vm_db
+
+    def process_alarm_for_vnf(self, policy):
+        '''call in plugin'''
+        vnf = policy['vnf']
+        params = policy['params']
+        mon_prop = policy['properties']
+        alarm_dict = dict()
+        alarm_dict['alarm_id'] = params['data'].get('alarm_id')
+        alarm_dict['status'] = params['data'].get('current')
+        driver = mon_prop['resize_compute']['event_type']['implementation']
+        return self.process_alarm(driver, vnf, alarm_dict)
+
+    def _invoke(self, driver, **kwargs):
+        method = inspect.stack()[1][3]
+        return self._alarm_monitor_manager.invoke(
+            driver, method, **kwargs)
+
+    def call_alarm_url(self, driver, vnf_dict, kwargs):
+        return self._invoke(driver,
+                            vnf=vnf_dict, kwargs=kwargs)
+
+    def process_alarm(self, driver, vnf_dict, kwargs):
+        return self._invoke(driver,
+                            vnf=vnf_dict, kwargs=kwargs)
+
+
 @six.add_metaclass(abc.ABCMeta)
 class ActionPolicy(object):
     @classmethod
@@ -269,29 +337,65 @@ class ActionRespawnHeat(ActionPolicy):
         vnf_id = vnf_dict['id']
         LOG.info(_('vnf %s dead and to be respawned'), vnf_id)
         if plugin._mark_vnf_dead(vnf_dict['id']):
-            plugin._vnf_monitor.mark_dead(vnf_dict['id'])
-            attributes = vnf_dict['attributes']
-            failure_count = int(attributes.get('failure_count', '0')) + 1
-            failure_count_str = str(failure_count)
-            attributes['failure_count'] = failure_count_str
-            attributes['dead_instance_id_' + failure_count_str] = vnf_dict[
-                'instance_id']
-            placement_attr = vnf_dict.get('placement_attr', {})
-            region_name = placement_attr.get('region_name')
-            # kill heat stack
-            heatclient = openstack.HeatClient(auth_attr=auth_attr,
-                                         region_name=region_name)
-            heatclient.delete(vnf_dict['instance_id'])
+            if vnf_dict['attributes'].get('monitoring_policy'):
+                plugin._vnf_monitor.mark_dead(vnf_dict['id'])
+                attributes = vnf_dict['attributes']
+                failure_count = int(attributes.get('failure_count', '0')) + 1
+                failure_count_str = str(failure_count)
+                attributes['failure_count'] = failure_count_str
+                attributes['dead_instance_id_' + failure_count_str] = vnf_dict[
+                    'instance_id']
+                placement_attr = vnf_dict.get('placement_attr', {})
+                region_name = placement_attr.get('region_name')
+                # kill heat stack
+                heatclient = openstack.HeatClient(auth_attr=auth_attr,
+                                                  region_name=region_name)
+                heatclient.delete(vnf_dict['instance_id'])
 
-            # TODO(anyone) set the current request ctxt instead of admin ctxt
-            context = t_context.get_admin_context()
-            _log_monitor_events(context, vnf_dict,
-                                "ActionRespawnHeat invoked")
-            update_vnf_dict = plugin.create_vnf_sync(context,
-                                                     vnf_dict)
-            LOG.info(_('respawned new vnf %s'), update_vnf_dict['id'])
-            plugin.config_vnf(context, update_vnf_dict)
-            plugin.add_vnf_to_monitor(update_vnf_dict, auth_attr)
+                # TODO(anyone) set the current request ctxt
+                context = t_context.get_admin_context()
+                _log_monitor_events(context, vnf_dict,
+                                    "ActionRespawnHeat invoked")
+
+                update_vnf_dict = plugin.create_vnf_sync(context,
+                                                         vnf_dict)
+                LOG.info(_('respawned new vnf %s'), update_vnf_dict['id'])
+                plugin.config_vnf(context, update_vnf_dict)
+                plugin.add_vnf_to_monitor(update_vnf_dict, auth_attr)
+
+            if vnf_dict['attributes'].get('alarm_url'):
+                attributes = vnf_dict['attributes']
+                failure_count = int(attributes.get('failure_count', '0')) + 1
+                failure_count_str = str(failure_count)
+                attributes['failure_count'] = failure_count_str
+                attributes['dead_instance_id_' + failure_count_str] = vnf_dict[
+                    'instance_id']
+                placement_attr = vnf_dict.get('placement_attr', {})
+                region_name = placement_attr.get('region_name')
+                # kill heat stack
+                heatclient = openstack.HeatClient(auth_attr=auth_attr,
+                                                  region_name=region_name)
+                heatclient.delete(vnf_dict['instance_id'])
+                vnf_dict['attributes'].pop('alarm_url')
+
+                # TODO(anyone) set the current request ctxt
+                context = t_context.get_admin_context()
+                _log_monitor_events(context, vnf_dict,
+                                    "ActionRespawnHeat invoked")
+                update_vnf_dict = plugin.create_vnf_sync(context,
+                                                         vnf_dict)
+                plugin.config_vnf(context, update_vnf_dict)
+
+
+@ActionPolicy.register('scaling')
+class ActionAutoscalingHeat(ActionPolicy):
+    @classmethod
+    def execute_action(cls, plugin, vnf_dict, scale):
+        vnf_id = vnf_dict['id']
+        plugin.create_vnf_scale(t_context.get_admin_context(), vnf_id, scale)
+        _log_monitor_events(t_context.get_admin_context(),
+                            vnf_dict,
+                            "ActionAutoscalingHeat invoked")
 
 
 @ActionPolicy.register('log')
@@ -314,6 +418,7 @@ class ActionLogAndKill(ActionPolicy):
                             "ActionLogAndKill invoked")
         vnf_id = vnf_dict['id']
         if plugin._mark_vnf_dead(vnf_dict['id']):
-            plugin._vnf_monitor.mark_dead(vnf_dict['id'])
+            if vnf_dict['attributes'].get('monitoring_policy'):
+                plugin._vnf_monitor.mark_dead(vnf_dict['id'])
             plugin.delete_vnf(t_context.get_admin_context(), vnf_id)
         LOG.error(_('vnf %s dead'), vnf_id)
